Add initial class layout and tests
This commits defines 2 primary data types
volume and storage pool

storage pools manage the allocation and lifecycle
of volumes using a storage backend.

a volume is a handel to a region of sotrage
aquried form a backend via a storage pool

a backend is an system of data storage with a
limited api  that can allocate volumes of specific sizes

this change provides a single in memory backend for testing.